update `LinearProgress` styling
